requirements.txt
asgiref==3.9.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cryptography==41.0.7
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.6
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.10
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
requests==2.31.0
sqlparse==0.5.3
typing_extensions==4.10.0
urllib3==2.0.7
